+++ server.js
@@ -19,6 +19,7 @@ const webchatController = require('./controllers/webchatController.js');
 
 app.set('view engine', 'ejs');
 app.set('views', './views');
+
 app.use(express.static(path.join(__dirname, 'public')));
 
 require('./socket/webchat')(io);
